requirements.txt
asgiref==3.5.0
Django==4.0.3
django-markdownx==4.0.0b1
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
python-dotenv==0.20.0
sqlparse==0.4.2
zipp==3.8.0
